Added enum suite for Cinema so that there are varying ticket price based on suite. Also added blockbuster check for extra charge

+++ src/Booking.java
@@ -30,7 +30,7 @@ public class Booking {
 		return TID;
 	}
 	
-	public float getTotalAmount() {
+	public double getTotalAmount() {
 		return seats.size() * ticketType.getPrice();
 	}
